+++ packages/api/src/utils/jwt.js
@@ -1,4 +1,4 @@
-/** @type {Record<string, HmacImportParams>} */
+/** @type {Record<string, any>} */
 const algorithms = {
   HS256: {
     name: 'HMAC',
@@ -10,7 +10,7 @@ const algorithms = {
 
 // Adapted from https://chromium.googlesource.com/chromium/blink/+/master/LayoutTests/crypto/subtle/hmac/sign-verify.html
 var Base64URL = {
-  stringify: function (/** @type {Uint8Array} */ a) {
+  stringify: function(/** @type {Uint8Array} */ a) {
     // @ts-ignore
     var base64string = btoa(String.fromCharCode.apply(0, a))
     return base64string
@@ -18,11 +18,14 @@ var Base64URL = {
       .replace(/\+/g, '-')
       .replace(/\//g, '_')
   },
-  parse: function (/** @type {string} */ s) {
-    s = s.replace(/-/g, '+').replace(/_/g, '/').replace(/\s/g, '')
+  parse: function(/** @type {string} */ s) {
+    s = s
+      .replace(/-/g, '+')
+      .replace(/_/g, '/')
+      .replace(/\s/g, '')
     return new Uint8Array(
       // @ts-ignore
-      Array.prototype.map.call(atob(s), function (c) {
+      Array.prototype.map.call(atob(s), function(c) {
         return c.charCodeAt(0)
       })
     )
@@ -151,7 +154,10 @@ export async function signJWT(payload, secret, alg = 'HS256') {
  * @param {string} token
  */
 export function decodeJWT(token) {
-  var output = token.split('.')[1].replace(/-/g, '+').replace(/_/g, '/')
+  var output = token
+    .split('.')[1]
+    .replace(/-/g, '+')
+    .replace(/_/g, '/')
   switch (output.length % 4) {
     case 0:
       break