+++ packages/api/test/scripts/worker-globals.js
@@ -1,22 +1,8 @@
-import stores from './stores.js'
-
 export const SALT = 'supersecret'
 export const DEBUG = '*'
-export const DEALS = stores.deals
-export const USERS = stores.users
-export const NFTS = stores.nfts
-export const NFTS_IDX = stores.nftsIndex
-export const METRICS = stores.metrics
-export const PINS = stores.pins
-export const FOLLOWUPS = stores.followups
-export const PINATA_QUEUE = stores.pinataQueue
-export const PINATA_JWT = 'test'
 export const CLUSTER_API_URL = 'http://localhost:9094'
 export const CLUSTER_BASIC_AUTH_TOKEN = 'test'
-export const CLUSTER_IPFS_PROXY_API_URL = 'http://localhost:9095'
-export const CLUSTER_IPFS_PROXY_BASIC_AUTH_TOKEN = 'test'
 export const CLUSTER_SERVICE = ''
-export const CLUSTER_ADDRS = ''
 export const MAGIC_SECRET_KEY = 'test'
 export const ENV = 'test'
 export const SENTRY_DSN = 'https://test@test.ingest.sentry.io/0000000'