+++ packages/client/src/platform.web.js
@@ -1,3 +1,5 @@
+import { MemoryBlockStore } from 'ipfs-car/blockstore/memory'
+
 export const fetch = globalThis.fetch
 export const FormData = globalThis.FormData
 export const Headers = globalThis.Headers
@@ -6,3 +8,4 @@ export const Response = globalThis.Response
 export const Blob = globalThis.Blob
 export const File = globalThis.File
 export const ReadableStream = globalThis.ReadableStream
+export const Blockstore = MemoryBlockStore