+++ packages/database_v2/migrations/default/1636945682111_sql-ingest-processor/up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE erc721_token_ingestion_queue
+ADD last_processed TIMESTAMP WITH TIME ZONE;