+++ packages/database_v2/migrations/default/1636951724069_sql-ingest-processor-iterator/down.sql
@@ -1,0 +1,3 @@
+DROP FUNCTION IF EXISTS ingest_unprocessed_tokens();
+DROP FUNCTION IF EXISTS get_unprocessed_tokens_in_queue();
+DROP FUNCTION IF EXISTS nullify_last_processed_ingest_queue();