+++ packages/database_v2/migrations/default/1637256388437_batchify-ingest-processor/down.sql
@@ -1,0 +1,2 @@
+DROP FUNCTION IF EXISTS ingest_unprocessed_tokens();
+DROP FUNCTION IF EXISTS get_unprocessed_tokens_in_queue();