+++ packages/website/components/button.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react'
+
 import Link from 'next/link'
 import clsx from 'clsx'
-
 import countly from '../lib/countly'
 
 /**
@@ -46,7 +46,7 @@ export default function Button({
   ...props
 }) {
   const onClickHandler = useCallback(
-    (event) => {
+    event => {
       tracking &&
         countly.trackEvent(tracking.event || countly.events.CTA_LINK_CLICK, {
           ui: tracking.ui,