+++ packages/website/pages/_app.js
@@ -1,11 +1,14 @@
-import { useEffect } from 'react'
-import Router from 'next/router'
-import { ReactQueryDevtools } from 'react-query/devtools'
 import 'tachyons/css/tachyons.css'
 import '../styles/global.css'
+import '../styles/nft-storage.swagger-ui.css'
+
 import { QueryClient, QueryClientProvider } from 'react-query'
+
 import Layout from '../components/layout.js'
+import { ReactQueryDevtools } from 'react-query/devtools'
+import Router from 'next/router'
 import countly from '../lib/countly'
+import { useEffect } from 'react'
 
 const queryClient = new QueryClient({
   defaultOptions: { queries: { staleTime: 60 * 1000 } },
@@ -19,7 +22,7 @@ const queryClient = new QueryClient({
 export default function App({ Component, pageProps }) {
   useEffect(() => {
     countly.init()
-    Router.events.on('routeChangeComplete', (route) => {
+    Router.events.on('routeChangeComplete', route => {
       countly.trackPageView(route)
     })
   }, [])
@@ -27,7 +30,7 @@ export default function App({ Component, pageProps }) {
   return (
     <QueryClientProvider client={queryClient}>
       <Layout {...pageProps}>
-        {(props) => <Component {...pageProps} {...props} />}
+        {props => <Component {...pageProps} {...props} />}
       </Layout>
       <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>